Fix and improve robot failure example notebook.
The robot failure example notebook did not contain the output, rendering it
not so informative an example.

The notebook did not download the data before using it.

It did not use the data imputation settings for `extract_relevant_features()`, leading to a crash.

Added timings for the extraction functions, as it's important to know that the current
implementation is surprisingly time consuming.

Consolidated the example plots and put them first, as that is more meaningful for someone
viewing the example with no background.

Added a bit of narrative text to give readers a basic idea what's going on.