Compute quantiles only once
change_quantiles() is called with several ql and qh arguments.
Quantile computation is costly, it is faster to compute all quantiles only once.
For each data chunk, this commit computes all quantiles which are needed, and
put them in a _quantiles argument, which is used as a cache variable.  If it is
set, it is used to speed-up computations.  Otherwise, everything is computed as
before.
This cache variable is not passed to convert_to_output_format.