Add "autolag" parameter to augmented_dickey_fuller() (#612)
* Add "autolag" parameter to augmented_dickey_fuller()

The result of Augumented Dickey-Fuller test deeply depends on how we
choose the lag parameter. StatsModels' adfuller() function provides
several criteria for that selection (AIC, BIC and t-stats etc).

This makes it a parameter, and enables users to tweak the lag selection
according to their preference.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>

* Add test case for invalid attribute names to increase coverage

This pattern has not been tested and coverall is complaining about
it. This patch fixes it.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>

* Apply review comments on GitHub

This contains a series of patches that fix issues pointed out by
Nils Braun.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>