Update: psd_aggregated
Features based on power spectral density: "mean_frequency", "median_frequency", "peak_frequency", "mean_power". The scipy's periodogram is used to calculate the base signal for the features. 

This features are useful in frequency domain analisys for EMG signals. The main reference for this features are: [1] Phinyomark, A., Phukpattaranont, P., & Limsakul, C. (2012). Feature reduction and selection for EMG signal classification. Expert systems with applications, 39(8), 7420-7431.