Make Dask and Distributed optional dependencies
Dask was de facto optional already (see the guarded import in
tsfresh/feature_extraction/data.py), but before this commit it was
installed by default.

Distributed is imported only to support Dask, when the classes
ClusterDaskDistribution or LocalDaskDistribution are used, so it is not
needed if Dask is not imported.

To install tsfresh with those dependencies, one can do:

    pip install -e ".[dask]"

This is needed, for example, to run the test suite. The documentation
has been updated accordingly.