Add a comment about multi-version local testing
Adds a comment about how to run tests locally on multiple versions of python.

Ensures it is clear that it only works if particular version of python is installed on the users machine.

Recommends `pytest` as a possible way to manage multiple local versions of python.

Note the hyperlink out to `setup.cfg` is pinned to commit sha 1297c8ca5bd6f8f23b4de50e3f052fb4ec1307f8 instead of pinning to main branch.